// packages/application/src/problems/interfaces/IProblemUseCases.d.ts
export interface ChangeProblemDifficultyInput {
    problemId: string;
    teacherId: string;
    difficultyLevel: number;
}